requirements.txt
ag2==0.9.3
annotated-types==0.7.0
anyio==4.9.0
asyncer==0.0.8
Authlib==1.6.0
bleach==6.2.0
bokeh==3.7.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.4
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
linkify-it-py==2.0.3
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
narwhals==1.44.0
numpy==2.3.1
packaging==25.0
pandas==2.3.0
panel==1.7.1
param==2.2.1
pillow==11.2.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyviz_comms==3.0.6
pywin32==310
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
six==1.17.0
sniffio==1.3.1
sseclient==0.0.27
termcolor==3.1.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
tiktoken==0.9.0
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uc-micro-py==1.0.3
urllib3==2.5.0
webencodings==0.5.1
xyzservices==2025.4.0
chromadb==0.5.23
python-pptx==1.0.2
beautifulsoup4==4.12.3
scikit-learn==1.6.1
